+++ vite.config.js
@@ -1,0 +1,31 @@
+/// <reference types="vitest" />
+import { defineConfig } from 'vite'
+import { resolve } from 'path'
+import dts from 'vite-plugin-dts'
+
+export default defineConfig({
+  build: {
+    lib: {
+      entry: resolve(__dirname, 'src/index.ts'),
+      name: 'Keshi',
+      // the proper extensions will be added
+      fileName: 'index',
+    },
+    rollupOptions: {
+      // make sure to externalize deps that shouldn't be bundled
+      // into your library
+      // external: ['vue'],
+      output: {
+        // Provide global variables to use in the UMD build
+        // for externalized deps
+        globals: {
+          // vue: 'Vue',
+        },
+      },
+    },
+  },
+  plugins: [dts({ rollupTypes: true })],
+  test: {
+    watch: false,
+  },
+})