+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 3.0.0
+
+This is a simplification and re-write.
+
+- BREAKING: Only store promises and `.resolve()` always returns a promise.
+- BREAKING: `getValue` is alwaus required as the 2nd arg to `.resolve()` and always a function.
+- BREAKING: The default import is a class and should be instantiated as `new Keshi()` instead of `createCache`.
+- BREAKING: Rename `del` to `delete` to be more clear and aligned with a ES6 Map.
+
+This is more ergonomic for Promise based caches which is the main goal of this library. Though you can still cache plain values (you just have to await them).
+
+Since the cache is no longer serializable and can only be in-memory, there's no option to pass a custom cache.
+
+You can freely use v2 if you disagree with the decision/don't want to update your code.
+
 # 2.0.7
 
 - Make `createCache` options in TS def optional.